% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_data.r
\docType{data}
\name{annotation_data}
\alias{annotation_data}
\title{Annotation dataset
Dataset contains gene length for human GRCh37 and mouse GRCm38 genome and low variability genes.}
\format{An object of class \code{list} of length 3.}
\usage{
data(annotation_data)
}
\description{
Annotation dataset
Dataset contains gene length for human GRCh37 and mouse GRCm38 genome and low variability genes.
}
\examples{
data(annotation_data)
gene_len_GRCh37 <- annotation_data[["gene_len_GRCh37"]]
gene_len_GRCm38 <- annotation_data[["gene_len_GRCm38"]]
LV_gene_human <- annotation_data[["LV_gene_human"]]
}
\keyword{datasets}
