% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{dimension}
\alias{dimension}
\title{dimension}
\arguments{
\item{dimension}{The dimensions of word vector representations for every word}
}
\description{
The dimensions of word vector representations for every word
}
\details{
The dimensions of word vector representations for every word
}
\keyword{internal}
