% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{doc}
\alias{doc}
\title{doc}
\arguments{
\item{doc}{List of document. each document}
}
\description{
List of document. each document
}
\details{
List of document. each document
}
\keyword{internal}
