% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{document_vector}
\alias{document_vector}
\title{Calculate Document Vector (SCDV)}
\usage{
document_vector(doc, wtv)
}
\arguments{
\item{doc}{List of document. each document}

\item{wtv}{Word-topic vectors}
}
\description{
Calculate Document Vector (SCDV)
}
