% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDV_functions_latest.r
\name{estimate_drop_out}
\alias{estimate_drop_out}
\title{This function is used for estimating dropout probability.}
\usage{
estimate_drop_out(sc_data, sc_data_expect, gene_len_org, per_tile_beta = 6,
  per_tile_tau = 6, alpha_init = c(1, -1), spois_init = 0.1,
  beta_init = c(0.1, 0.1, 0.1), tau_init = c(0.1, -0.1, -0.1),
  em_error_par = 0.01, em_min_count = 1, em_max_count = 30,
  trace_flag = 0)
}
\arguments{
\item{sc_data}{Input count matrix.}

\item{sc_data_expect}{Gene expression from the pool neighboring cells obtained by function get_expected_cell}
}
\description{
This function is used for estimating dropout probability.
}
