% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{gmm_args}
\alias{gmm_args}
\title{gmm_args}
\arguments{
\item{gmm_args}{Parameters for GMM model ( parameters of ClusterR::center_scale and ClusterR::GMM )}

\item{args}{Parameters for GMM model ( parameters of ClusterR::center_scale and ClusterR::GMM )}
}
\description{
Parameters for GMM model ( parameters of ClusterR::center_scale and ClusterR::GMM )
}
\keyword{internal}
