% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDV_functions_latest.r
\name{qc_check}
\alias{qc_check}
\title{QC check}
\usage{
qc_check(data_count)
}
\arguments{
\item{data_count}{Input count matrix.}
}
\description{
This function allows you to check the library size and detaction rate.
}
\examples{
\dontrun{
   qc_check(input_data)
}
}
\keyword{qc}
