% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{scdv}
\alias{scdv}
\title{Calculate Sparse Composite Document Vector (SCDV)}
\usage{
scdv(doc, k, dimension, p = 0.01, word2vec_args = list(),
  gmm_args = list())
}
\arguments{
\item{doc}{List of document. each document}

\item{k}{The number of clusters}

\item{dimension}{The dimensions of word vector representations for every word}

\item{p}{The sparsity threshold for SCDV}

\item{word2vec_args}{Parameters for wrod2vec model ( parameters of wordVectors::train_word2vec )}

\item{gmm_args}{Parameters for GMM model ( parameters of ClusterR::center_scale and ClusterR::GMM )}
}
\description{
Calculate Sparse Composite Document Vector (SCDV)
}
