% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDV_functions_latest.r
\name{scdv_main}
\alias{scdv_main}
\title{Differential hyper-variability test}
\usage{
scdv_main(treatment_data, treatment_data_weight, control_data,
  control_data_weight, num_permute = 1000, span_param = 0.5, ncore = 1)
}
\arguments{
\item{treatment_data}{Normalized count data for the treatment group}

\item{treatment_data_weight}{1 - dropout probability for the treatment group}

\item{control_data}{Normalized count data for the control group}

\item{control_data_weight}{1 - dropout probability for the control group}

\item{num_permute}{Number of permutation performed in the test}

\item{span_param}{The span parameter in loess when fitting the mean-variance curve}

\item{ncore}{Number of CPU cores used in the test}
}
\value{
\item{sf_treatment}{The hyper-variability statistics for the treatment group}
 \item{sf_control}{The hyper-variability statistics for the control group}
 \item{sf_diff}{Difference between sf_treatment and sf_control: sf_treatment - sf_control}
 \item{sf_diff_pval}{P-values for testing whether the hyper-variability of each gene in the treamtment group is larger than that in the control group}
 \item{sf_diff_fdr}{Adjusted p-values (FDR) of sf_diff_pval}
 \item{sf_diff_pval_alt}{P-values for testing whether the hyper-variability of each gene in the treamtment group is smaller than that in the control group}
 \item{sf_diff_fdr_alt}{Adjusted p-values (FDR) of sf_diff_pval_alt}
 \item{sf_diff_pval_ts}{P-values for testing whether the hyper-variability of each gene in the treamtment group is not equal to that in the control group}
 \item{sf_diff_fdr_ts}{Adjusted p-values (FDR) of sf_diff_pval_ts}
}
\description{
This function is used for testing differential hyper-variability.
}
\examples{
\dontrun{
diff_disper <- scdv_main(treatment_data_adjust,treatment_data_weight,control_data_adjust,control_data_weight,num_permute=10000,span_param=0.5,ncore=6)
write.csv(cbind(match_gene_name,diff_disper),file="diff_hypervar.csv",row.names=FALSE)
}
}
\keyword{differential}
\keyword{hyper-variability}
\keyword{test}
