% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDV_functions_latest.r
\name{test_mean_main}
\alias{test_mean_main}
\title{Differential mean test}
\usage{
test_mean_main(treatment_data, treatment_data_weight, control_data,
  control_data_weight, num_permute = 1000, ncore = 1)
}
\arguments{
\item{treatment_data}{Normalized count data for the treatment group}

\item{treatment_data_weight}{1 - dropout probability for the treatment group}

\item{control_data}{Normalized count data for the control group}

\item{control_data_weight}{1 - dropout probability for the control group}

\item{num_permute}{Number of permutation performed in the test}

\item{ncore}{Number of CPU cores used in the test}
}
\value{
\item{statistics}{The weighted t-statistics}
 \item{pval_greater}{P-values for testing whether the expression of each gene in the treatment group is larger than that in the control group}
 \item{fdr_greater}{Adjusted p-values (FDR) of pval_greater}
 \item{pval_less}{P-values for testing whether the expression of each gene in the treatment group is smaller than that in the control group}
 \item{fdr_less}{Adjusted p-values (FDR) of pval_less}
 \item{pval_ts}{P-values for testing whether the expression of each gene in the treatment group is not equal to that in the control group}
 \item{fdr_ts}{Adjusted p-values (FDR) of pval_ts}
}
\description{
This function is used for testing differential gene expression.
}
\examples{
\dontrun{
diff_expr <- test_mean_main(treatment_data_adjust,treatment_data_weight,control_data_adjust,control_data_weight,num_permute=10000,ncore=6)
write.csv(data.frame(match_gene_name,diff_expr),file="diff_expr.csv",row.names=FALSE)
}
}
\keyword{differential}
\keyword{mean}
\keyword{test}
