% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{word2vec}
\alias{word2vec}
\title{Get word embedding expression by word2vec}
\usage{
word2vec(doc, dimension, args = list())
}
\arguments{
\item{doc}{List of document. each document}

\item{dimension}{The dimensions of word vector representations for every word}

\item{args}{Parameters for wrod2vec model ( parameters of wordVectors::train_word2vec )}
}
\description{
Get word embedding expression by word2vec
}
