% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{word2vec_args}
\alias{word2vec_args}
\title{word2vec_args}
\arguments{
\item{word2vec_args}{Parameters for wrod2vec model ( parameters of wordVectors::train_word2vec )}

\item{args}{Parameters for wrod2vec model ( parameters of wordVectors::train_word2vec )}
}
\description{
Parameters for wrod2vec model ( parameters of wordVectors::train_word2vec )
}
\details{
Parameters for wrod2vec model ( parameters of wordVectors::train_word2vec )
}
\keyword{internal}
