% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdv.R
\name{word_topic_vector}
\alias{word_topic_vector}
\title{Calculate Word-Topic Vector}
\usage{
word_topic_vector(doc, k, dimension, word2vec_args = list(),
  gmm_args = list())
}
\arguments{
\item{doc}{List of document. each document}

\item{k}{The number of clusters}

\item{dimension}{The dimensions of word vector representations for every word}

\item{word2vec_args}{Parameters for wrod2vec model ( parameters of wordVectors::train_word2vec )}

\item{gmm_args}{Parameters for GMM model ( parameters of ClusterR::center_scale and ClusterR::GMM )}
}
\description{
Calculate Word-Topic Vector
}
